fix(front): модель вынесена в entities
