+++ frontend/src/entities/tasks/model/index.ts
@@ -1,1 +1,3 @@
 export { useTaskStore as taskModel } from "./tasks";
+export type {Task} from "./models"
+export type {Status} from "./models"
