--- frontend/src/shared/api/typicode/models.ts
@@ -1,32 +1,0 @@
-export interface Task {
-  id: string;
-  url: string;
-  type: {
-    id: number;
-    name: string;
-    description: string;
-  };
-  name: string;
-  status: Status;
-  statusId: number;
-  // parserTaskWebsiteTagsOptions: {
-  //   id: string,
-  //   parserTaskWebsiteTags: Array<{
-  //     id: string,
-  //     findOptions : {
-  //       name: string,
-  //       atributs: Array<{
-  //         name: string,
-  //         value: string
-  //       }>
-  //     }
-
-  //   }>
-  // }
-}
-
-export interface Status {
-  id: number;
-  key: string;
-  description: string;
-}
